tcrun: move to delayed iteration (caching iteration)
This allows us to avoid the need for immediately iterating the entire
directory tree and enumerating every installation. Instead, we delay the
enumeration until we need the specific value (either to search or
iterate). The desire is to manage to get to the point where we can avoid
the iteration for most of the cases.